move oncokb content to AnnotationHeader.tsx
