+++ src/pages/patientView/copyNumberAlterations/column/AnnotationColumnFormatter.tsx
@@ -26,7 +26,7 @@ import { IAnnotationColumnProps } from 'shared/components/mutationTable/column/A
 import { CancerGene, IndicatorQueryResp } from 'oncokb-ts-api-client';
 import { getAlterationString } from 'shared/lib/CopyNumberUtils';
 import { getCivicCNAVariants } from 'shared/lib/CivicUtils';
-import { AnnotationHeader } from 'shared/components/mutationTable/column/annotation/AnnotationHeaderTooltipCard';
+import AnnotationHeader from 'shared/components/mutationTable/column/annotation/AnnotationHeader';
 
 /**
  * @author Selcuk Onur Sumer
